refactor: Splits up giant client file
